actually install zipped parsing
this refactors the proxy and the parser a teeny tiny bit, but
bare with me.

Instead of using strings to communicate the response body between proxy
and parser, we use byte arrays that are sent via base64 encoded strings
this allows zipped and unzipped data to be recorded.

Note: We still unzip the response and use the unzipped value for
comparison in tests